refactor: replace alertLogin with ModalLogin component
Alert 시스템에서 독립적인 모달 컴포넌트로 로그인 UI 분리.
module context를 활용한 Promise 기반 requestLogin() 함수로 변경.
