+++ src/lib/Setting/Pages/UserSettings.svelte
@@ -8,7 +8,7 @@
     import { alertConfirm} from "src/ts/alert.svelte";
     import { forageStorage, loadInternalBackup } from "src/ts/globalApi.svelte";
     import { isTauri, isNodeServer, isCapacitor } from "src/ts/platform"
-    import { unMigrationAccount } from "src/ts/storage/accountStorage";
+    import { unMigrationAccount } from "src/ts/storage/accountStorage.svelte";
     import { checkDriver } from "src/ts/drive/drive.svelte";
     import { LoadLocalBackup, SaveLocalBackup } from "src/ts/drive/backuplocal";
     import Button from "src/lib/UI/GUI/Button.svelte";
