+++ src/ts/characterCards.svelte.ts
@@ -24,7 +24,7 @@ import { CharXImporter, CharXSkippableChecker, CharXWriter } from "./process/pro
 import { exportModule, readModule, type RisuModule } from "./process/modules"
 import { readFile } from "@tauri-apps/plugin-fs"
 import { onOpenUrl } from '@tauri-apps/plugin-deep-link';
-import { AccountStorage } from "./storage/accountStorage"
+import { AccountStorage } from "./storage/accountStorage.svelte"
 
 
 const EXTERNAL_HUB_URL = 'https://sv.risuai.xyz';
